refactor(web-serial-data-access): command pipelineに実行責務を統合
CommandQueue/CommandRunner/CommandService の実行系を SerialCommandPipelineService に集約し、receive$ ベースのバッファと直列キュー/キャンセル制御を単一クラスで扱えるようにする。
