--- libs/web-serial/data-access/src/lib/serial-command/serial-command-queue.service.ts
@@ -1,168 +1,0 @@
-/** Full rewrite (#606). Serializes command jobs before runner hits the port. */
-import { Injectable } from '@angular/core';
-import {
-  EMPTY,
-  Observable,
-  Subject,
-  Subscriber,
-  catchError,
-  concatMap,
-  defer,
-  finalize,
-  mergeMap,
-  throwError,
-} from 'rxjs';
-
-/** {@link SerialCommandQueueService#enqueueCommand$} のキュー制御オプション（issue #565） */
-export interface SerialCommandEnqueueOptions {
-  /**
-   * `true` のとき、この enqueue の直前に **まだ実行が始まっていない** 先行ジョブを棄却する。
-   * 実行中のジョブは止めない。{@link SerialCommandQueueService#cancelAllCommands} とは別。
-   */
-  cancelPrevious?: boolean;
-}
-
-/**
- * {@link SerialCommandQueueService} の内部キュー状態（issue #648）。
- *
- * - **generation**: `cancelAllCommands` ごとに増える。各 enqueue はその時点の世代を覚え、
- *   `defer` 実行直前に世代が変わっていればジョブ全体を棄却する（`cancelAllCommands` 経路）。
- * - **pendingCount**: キューに載せたが `finalize` まで完了していないジョブ数（`getPendingCommandCount`）。
- * - **nextSlotId**: enqueue のたびに単調増加する ID。実行順とは一致しない。
- * - **rejectPendingSlotsBelow**: `cancelPrevious`（`invalidatePendingEnqueues`）で設定されるしきい値。
- *   `defer` 内で `assignedSlotId` がこの値未満なら未実行扱いとして棄却する。`cancelAllCommands` で `null` に戻す。
- */
-export interface CommandQueueState {
-  /** `cancelAllCommands` のたびに増加。enqueue 時点の世代と比較して実行可否を決める */
-  generation: number;
-  pendingCount: number;
-  /** enqueue ごとに増える単調 ID（実行順と一致しない）。`cancelPrevious` 時の棄却判定に使う */
-  nextSlotId: number;
-  /**
-   * 実行開始時、`assignedSlotId` がこの値未満なら棄却（`cancelPrevious` で設定）。
-   * `cancelAllCommands` ではリセットする。
-   */
-  rejectPendingSlotsBelow: number | null;
-}
-
-/**
- * シリアルコマンド実行を直列化し、世代ベースで一括キャンセルする。
- *
- * **契約**
- * - `enqueueCommand$` 1 回につき、呼び出し元の Observable は **単一の next の後 complete**（または error）する。
- * - 内部の `concatMap` により、前のジョブが完了するまで次の `factory` は走らない（直列実行）。
- * - `factory` は `defer` 内で実行される。実行直前に `state.generation` を再チェックし、
- *   `cancelAllCommands` 済みなら `All commands cancelled` で打ち切る。
- * - `enqueuedGen` は enqueue 時点の世代。キャンセル後もキューに積まれた古いジョブは
- *   実行開始時に世代不一致で棄却される。
- * - **スロット**（`assignedSlotId`）は `cancelPrevious` 時に未実行ジョブだけを棄却するために使う。
- */
-@Injectable({
-  providedIn: 'root',
-})
-export class SerialCommandQueueService {
-  private readonly executionQueue$ = new Subject<Observable<unknown>>();
-  private readonly state: CommandQueueState = {
-    generation: 0,
-    pendingCount: 0,
-    nextSlotId: 0,
-    rejectPendingSlotsBelow: null,
-  };
-
-  constructor() {
-    this.executionQueue$
-      .pipe(
-        concatMap((work) =>
-          work.pipe(
-            catchError((err: unknown) => {
-              console.error('Serial command queue work error:', err);
-              return EMPTY;
-            }),
-          ),
-        ),
-      )
-      .subscribe();
-  }
-
-  /**
-   * キューに載っているがまだ `factory` が起動していないジョブを棄却するしきい値を更新する。
-   * 直後に enqueue されるコマンドは新しいスロット ID を得る。
-   */
-  invalidatePendingEnqueues(): void {
-    this.state.rejectPendingSlotsBelow = this.state.nextSlotId + 1;
-  }
-
-  /**
-   * enqueue 時点の世代がまだ有効か（cancel されていなければ true）
-   */
-  isGenerationActive(enqueuedGen: number): boolean {
-    return this.state.generation === enqueuedGen;
-  }
-
-  enqueueCommand$<T>(
-    factory: (enqueuedGen: number) => Observable<T>,
-    opts?: SerialCommandEnqueueOptions,
-  ): Observable<T> {
-    return new Observable<T>((subscriber) => {
-      if (opts?.cancelPrevious) {
-        this.invalidatePendingEnqueues();
-      }
-      const assignedSlotId = ++this.state.nextSlotId;
-      const enqueuedGen = this.state.generation;
-      this.state.pendingCount++;
-      this.executionQueue$.next(
-        this.createQueuedWork$(
-          enqueuedGen,
-          assignedSlotId,
-          factory,
-          subscriber,
-        ),
-      );
-    });
-  }
-
-  /**
-   * concatMap 1 単位: 世代チェック → factory 実行 → 結果を呼び出し元 subscriber へ一度だけ配送
-   */
-  private createQueuedWork$<T>(
-    enqueuedGen: number,
-    assignedSlotId: number,
-    factory: (enqueuedGen: number) => Observable<T>,
-    subscriber: Subscriber<T>,
-  ): Observable<unknown> {
-    return defer(() => {
-      if (this.state.generation !== enqueuedGen) {
-        return throwError(() => new Error('All commands cancelled'));
-      }
-      if (
-        this.state.rejectPendingSlotsBelow !== null &&
-        assignedSlotId < this.state.rejectPendingSlotsBelow
-      ) {
-        return throwError(() => new Error('All commands cancelled'));
-      }
-      return factory(enqueuedGen);
-    }).pipe(
-      finalize(() => {
-        this.state.pendingCount--;
-      }),
-      mergeMap((value) => {
-        subscriber.next(value as T);
-        subscriber.complete();
-        return EMPTY;
-      }),
-      catchError((err: unknown) => {
-        subscriber.error(err);
-        return EMPTY;
-      }),
-    );
-  }
-
-  cancelAllCommands(): void {
-    this.state.generation++;
-    this.state.rejectPendingSlotsBelow = null;
-  }
-
-  getPendingCommandCount(): number {
-    return this.state.pendingCount;
-  }
-}
